remove Makefile (Closes: #21)
the makefile doesn't bring any advantage over the normal `python
setup.py install` instructions and just add an extra
dependencies. furthermore, it installs files in a non-standard
location that should generally be avoided.

it will also break Debian packaging, which will think this is a C
program and try to install it as such (and then in the wrong location).